Don't do workaround if UEFI_PFLASH_VARS is present
With the QEMU backend refactor, three new variables have been introduced for
configuring UEFI firmware on pflash drives:

UEFI_PFLASH_CODE - Specifies the file containing the read-only UEFI firmware
code.
UEFI_PFLASH_VARS - Specifies the file which contains the UEFI firmware
variables.
PUBLISH_PFLASH_VARS - Specifies the name of the file to publish the pflash
variables to which can be used in a chained test.

If the VARS file is present then there is no need to perform the aarch64
workaround because the VARS file should contain the correct device to boot
from.