Use MIRROR_HTTP as main repo in Kubic
This PR disables all the default repositories (including
the ones from download.opensuse.org and the DVD device) that
are present after the DVD installation of Kubic.

Then it adds the *MIRROR_HTTP* as the only available repo,
which is used for testing purposes. If this variable is not
present, the test will fail.